more PR comments
- removed internal check::prelude
- removed some unused imports found using `cargo udeps`
- other tweaks
