+++ aziot/Cargo.toml
@@ -8,14 +8,12 @@ edition = "2018"
 [dependencies]
 anyhow = "1.0.34"
 async-trait = "0.1.42"
-backtrace = "0.3"
 base64 = "0.12"
 byte-unit = "4.0.9"
 chrono = { version = "0.4", features = ["serde"] }
 colored = "2.0.0"
 derive_more = "0.99.11"
 erased-serde = "0.3.12"
-futures = "0.3"
 hyper = "0.13"
 hyper-openssl = "0.8"
 nix = "0.18"
