+++ aziot/aziot-check-common/Cargo.toml
@@ -6,6 +6,5 @@ edition = "2018"
 
 
 [dependencies]
-anyhow = "1.0.34"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0.59"
