+++ aziot/src/internal/check/checks/hostname.rs
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-use super::prelude::*;
+use anyhow::{anyhow, Result};
+use serde::Serialize;
+
+use crate::internal::check::{CheckResult, Checker, CheckerCache, CheckerMeta, CheckerShared};
 
 #[derive(Serialize, Default)]
 pub struct Hostname {
