+++ aziot/src/internal/check/checks/well_formed_configs.rs
@@ -1,11 +1,14 @@
 // Copyright (c) Microsoft. All rights reserved.
 
-use super::prelude::*;
-
-use std::path::Path;
-
+use anyhow::{Context, Error, Result};
+use serde::Serialize;
 use tokio::fs;
 use tokio::io;
+use tokio::io::AsyncReadExt;
+
+use crate::internal::check::{CheckResult, Checker, CheckerCache, CheckerMeta, CheckerShared};
+
+use std::path::Path;
 
 pub fn well_formed_configs() -> impl Iterator<Item = Box<dyn Checker>> {
     let mut v: Vec<Box<dyn Checker>> = Vec::new();
