+++ http-common/Cargo.toml
@@ -10,7 +10,6 @@ base64 = "0.12"
 futures-util = "0.3"
 http = "0.2"
 hyper = { version = "0.13", optional = true }
-lazy_static = "1"
 libc = "0.2"
 log = "0.4"
 nix = "0.18"
