+++ identity/aziot-dps-client-async/Cargo.toml
@@ -24,10 +24,8 @@ url = "2"
 
 aziot-cert-client-async = { path = "../../cert/aziot-cert-client-async" }
 aziot-cloud-client-async-common = { path = "../aziot-cloud-client-async-common" }
-aziot-key-client = { path = "../../key/aziot-key-client" }
 aziot-key-client-async = { path = "../../key/aziot-key-client-async" }
 aziot-key-common = { path = "../../key/aziot-key-common" }
-aziot-key-openssl-engine = { path = "../../key/aziot-key-openssl-engine" }
 aziot-tpm-client-async = { path = "../../tpm/aziot-tpm-client-async" }
 aziot-tpm-common = { path = "../../tpm/aziot-tpm-common" }
 http-common = { path = "../../http-common" }
