+++ identity/aziot-identity-common-http/Cargo.toml
@@ -13,5 +13,4 @@ serde_json = "1"
 aziot-cert-common-http = { path = "../../cert/aziot-cert-common-http" }
 aziot-identity-common = { path = "../aziot-identity-common" }
 aziot-key-common = { path = "../../key/aziot-key-common" }
-aziot-key-common-http = { path = "../../key/aziot-key-common-http" }
 http-common = { path = "../../http-common" }
